Feature/extended querying

+++ src/Tributech.DataSpace.TwinAPI/Model/PaginatedResponse.cs
@@ -3,12 +3,12 @@
 
 namespace Tributech.DataSpace.TwinAPI.Model {
 	public class PaginatedResponse<T> where T : class {
-		public PaginatedResponse(int totalElements, IEnumerable<T> content) {
+		public PaginatedResponse(long totalElements, IEnumerable<T> content) {
 			TotalElements = totalElements;
 			Content = content;
 		}
 
-		public int TotalElements { get; }
+		public long TotalElements { get; }
 		public IEnumerable<T> Content { get; }
 
 	}
